// front/src/api/model/blogPost.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * FastAPI React Backend
 * OpenAPI spec version: 1.0.0
 */
import type { BlogPostUpdatedAt } from './blogPostUpdatedAt';

/**
 * ブログ記事レスポンスモデル
 */
export interface BlogPost {
  title: string;
  content: string;
  author: string;
  id: string;
  created_at: string;
  updated_at?: BlogPostUpdatedAt;
}